requirements.txt
asgiref==3.4.1
attrs==21.3.0
bcrypt==3.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
distro==1.6.0
Django==3.2.7
django-cpf==0.1.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
paramiko==2.9.1
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
texttable==1.6.4
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==0.59.0
zipp==3.6.0
